Add support for pulling images from Github
This adds a

  capstan pull <username>/<repository>

command for pulling base images from Github.  You can now type:

  capstan pull penberg/osv-base

to pull my base image and use that in the Capstanfile to avoid compiling
OSv altogether.

Please note that this breaks backwards compatibility because repository
layout is changed to support git.  Images are no longer stored directly
in the repository root but put in their own directory so that we can
just use 'git clone' for pulling images.

And yes, this needs lots more work.  For example:

  - support for arbitrary git repositorys, not just Github

  - use 'git pull' to update the base image

  - use git tags for versioning

  - make 'capstan push' work with git

  - etc, etc.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>