RPM support
Add 'rpm-base' definition to Capstanfile that makes the build command
use Fedora RPM as the base for an image in addition to the OSv base
image.  The files are uploaded to an image with the rpm2cpio command.

Note that we're using a hard-coded Fedora HTTP repository URL and don't
deal with dependencies yet.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>