run: Fix broken image detection logic
I tried running Don's HTTP server example but it 'capstan run' fails:

[penberg@localhost http-server]$ capstan run
http-server: image format not recognized, unable to run it.

The problem is that there's an ELF executable with the same
'http-server' name in the directory:

  [penberg@localhost http-server]$ file http-server
  http-server: ELF 64-bit LSB executable, x86-64, version 1 (SYSV),
  dynamically linked (uses shared libs), for GNU/Linux 2.6.32,
  BuildID[sha1]=0xa72de03d76d08142358c11686250d12202ac4c37, not stripped

which the run command attempts to run.  Fix the problem by looking up
the default image from local repository, not from the current directory.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>