Capstanfile basename substitution
This patch adds basename variable substitution to simplify file listing
in Capstanfile.

You can now do this to specify target and source file mapping:

  /usr/app/app.jar: build/&

which expands to:

  /usr/app/app.jar: build/app.jar

where "/usr/app/app.jar" is the path in the built image and
"build/app.jar" is the file that is picked up from the local filesystem.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>