cmd/run: Fix capstan run
1) capstan run
Capstanfile must be present under pwd.
It will build the image and create an instance with the name specified
in Capstanfile filed name (Now, it is the parent directory name)
If the instance already exists, it will be deleted and a new instance
with the same name will be created.

2) capstan run $instance_name
It will restart an stopped instance. The instance must exist.

3) capstan run -i $image_name $instance_name
It will create an instance using the image. The image must be present.
If the instance_name already exist it will be deleted and a new instance
with the same name will be created.

Signed-off-by: Asias He <asias@cloudius-systems.com>