pull: add proxy support
In proxy environment, we can use capstan search to list all images while
we can't use capstan pull to download it. We found that the default http
client uses a default tansport that supports proxy. But the transport of
DisableCompression flag has covered the  default transport, so we may
add the proxy ability.

Signed-off-by: wangbicheng <wangbicheng@huawei.com>
Signed-off-by: Pekka Enberg <penberg@iki.fi>