build: Fix build command - discard stdout and stderr in non-verbose mode
I am running on Ubuntu 14.10 and I noticed that on my computer the
command 'capstan build' hangs during the upload files stage when in
normal (i.e.  non verbose) mode, while everything works fine when the
files are uploaded in verbose mode.

After some troubleshooting, I came to the conclusion that the problem is
due to the fact that in the file 'cmd/build.go' when the subprocess for
'qemu-system-x86_64' (i.e. the process that writes the file to the OSv
image being built) is forked, the sdtout and stderr descriptors for that
process are copied to the OS stdout and stderr only in verbose mode, and
they are not discarded (as they should) in the normal (non verbose)
case.  This can cause the subprocess (and hence the whole build command)
to hang because it is trying to write all the 'Adding <filename>'
messages to stdout/stderr and those pipes are not set up.

I think the correct approach would be to discard (i.e. send to
'/dev/null') those two file descriptors in this case.

I created the patch below that addresses this problem and with this
patch 'capstan build' runs to completion and does not hang on my
computer.

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>