Add NBD file for manipulation of images
Current implementation of the build command was specific to one use case
required by the build command. With this commit we are introducing a new type
for updating image files through QEMU's command qemu-nbd. The functionality
follows the one in OSv's scripts/imgedit.py supporting reading and writing of
image sections. The main reason for this change is that we are introducing a
new command simplifying the composition of images which, to some extent,
duplicates the build command, however it also requires other image-related
functionalities.

The nbd.File object provides all the functionalities required to manipulate the
content of the image file. It provides a high level API for writing to the file
as if it were an ordinary file.

Two helper functions have also been added, SetPartition and SetCmdLine for
initialising the partition table and setting the command line of the virtual
image, respecitevly.

The build command has been rewritten to use the new nbd.File.

Important notice: this patch is temporal solution. In the future we should
remove the dependency on qemu-nbd (and QEMU) completely to allow cross platform
composition of images and packages.

Change-Id: Ia99d6e18bb285a4a1ac4ff61260cc60ebefe2272
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>