Introduce the compose command for Capstan
The 'capstan compose' command enables composition of QEMU images from file or
folder. Images are based on an empty OSv loader images containing only the kernel
and the mkfs and cpiod tools to create the additional user ZFS partition and
copy files into this partition.

Current implementation assumes there is a mike/osv-launcher image available in
the repository. The name was chosen to prevent potential conflicts with other
images, e.g., cloudus/osv-base. The user is allowed to configure the size of
the user partition which should not be less than 64 MB due to a constraint in
the mkfs tool. As opposed to OSv's build scripts where the size of the image is
the size of the entire VM, Capstan's compose uses the specified size only for
the user parition, not including the size of loader so that the user's
partition will be exactly what the user requested.

Besides setting the name of the image, the user has to provide a path. This
can either be a file or a directory. In the latter case, the directory is
uploaded into the image recursively.

The compose command does not allow users to set the command line at the moment.
Thus, when launching the VM based on such an image the user must specify the
command to be executed.

Future versions will include automatic dependency retrieval (for example, the
user will be able to provide a path to a binary which will look for all
dependent libraries and upload them onto the image) and specification of
additional packages, for example cli, HTTP server, ...

Change-Id: I28ad31f0343ddf4d1aa5fdf8971b5e6c315a71bf
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>