Package initialization command
This patch introduces the first in line of package related commands:
initialisation of the package structure, compliant with the MIKELANGELO package
meta data. It is basically a simple YAML consisting of name, author, version
fields. Name and author are mandatory fields.

Optionally, the package structure also asumes the use of dependencies and
provided binaries through two additional fields, namely require and binary. The
former allows referencing other packages while composing OSv-based image and
the latter specifies which binaries should be provided as convenience methods
for the end user.

Change-Id: Ia5f75e5c703010facadee48d50fbfb8f860b1a17
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>