Reorganise the NBD utilities
To facilitate integration of an MPM package, a majority of code used to compose
images from plain files/directories needs to be re-used. This patch prepares
the underlying functionality and updates the existing "compose" command to use
it appropriately.

The main change in this patch is that we have moved NBD from a dedicated
package to util package. The reason for this was that NBD package was somehow
abused. To prevent circular imports caused by using the util package from nbd
we are moving NBD-related functionality to a util package where it will
reside until we remove the dependency on the NBD file completely. As we have
already agreed in previos patches, NBD (and QEMU for that matter) should not be
required while packaging applications for OSv.

Change-Id: I889353e286e0d8b6db0645ba04a416e3ff94faed
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>