Introduce the package compose command
This is the first in series of patches that are going to support package
composition of virtual images from MPM packages.  An MPM package consists of a
metadata package descriptor and the directories and files that must be uploaded
onto the virtual image.

We start with the package descriptor parsing and collection of files required
for upload.

Change-Id: I69714f858521b3d841f89c4367dc580e29cc2c1d
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>