Compose packages using required dependencies
This patch builds on the previous adding all the required packages into the
target virtual image. If any of the required packages is missing, the process
is stopped as there is no way to have an optional dependency. Dependencies are
specified in meta/package.yaml file with an optional list "require".

The capstan compose command has also been updated to use the common mechanism
for uploading files with capstan package compose.

Packages are stored within Capstan's directory (~/.capstan/packages). To this
end we have updated the way repository constructs paths for images and
packages. Tests have been added. Prior to using packages, they must be
(currently) manualy added to the repository.

Change-Id: Ieff933849bf6051f3ff3f114382cb5fa61b2ea85
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>