Fix repository path setting
In one of my previous commits I have changed the way the base path of the
repository is set because I needed the same base path for storing the packages.
This broke all existing operations on the repository.

This patch fixes the problem and uses the correct path to the image/instance
repository.

Change-Id: Iae5699cc1b9c89bf264544f8d5eca38189a54f04
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>