Extend package manifest with Title attribute
Previously, we identified the packages using Name attribute which was
neither completely human readable nor machine identifiable. This patch
introduces an additional attribute (Title) that allows the package
maintainer to handle the title of the package (human readable) and
the package name or identifier (Name). The latter is also used to
reference dependencies.

This patch furthermore adds an utility parsing function that helps
retrieving package manifest from a file.

Change-Id: I1409a9bd1b02e7095af1e7683a95716ea19bf93e
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>