Replace package directories with package files
The patch introduces the application package as an "mpm" file
containing all the resources and the package manifest. Essentially,
this is a tar file with a specific extension. Even though the mpm file
also contains the manifest, it is also stored in Capstan's repository
for more efficient retrieval of the packages and their dependencies.

Two additional commands are furthermore introduced: build and import.
The former will take the contents of the current directory and build
a package out of it (if it also contains the package manifest). The
import command will build the package and import it into Capstan's
local repository from where it can be used to compose application
images.

Change-Id: Iff6f62e3e753e98cee188c413abcc357519adbc9
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>