Include "run" switch to the package compose command
Using the "run" switch one can provide the default command line to be
executed when the VM is started. If cmd is not provided, an empty
default command is set meaning that the VM will start but will
immediately terminate, even though if there are commands executed in the
background.

Change-Id: Ia07258c2a2f8e4aff6ea967982f4d40d4c610239
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>