Enable symbolic links in MPM packages
Only regular files and directories have been allowed in previos version
of MPM. Using a package with symbolic links the collection, composition
and package building would fail either with an error or not including
symlinks in the corresponding results.

This patch addresses the issue of symbolic links preserving them in the
original form, i.e. they remain relative. This also means that symbolic
links outside of the package sandbox are not supported. In the future we
may need to add such links, however this would then require that the
symbolic links is replaced with the actual copy of the target in order
to allow packages to be portable.

During the package extraction, symbolic links are also properly restored
from the archive.

Change-Id: I424acc61bf4e69395cf81fb07999fa00fc65e1f2
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>