Package overrides the content of dependent packages
Previously, it was impossible to replace any of the files that have been
contained in any of the dependent files. For example, it was not
possible to replace the /etc/hosts file with a predefined list of known
hosts.

This patch fixes this by swapping the order of package composition. The
process now starts with dependent packages allowing the package being
built to override any of the files already provided by dependent
packages.

Change-Id: I7cac59828c3558524b80b484f69f144c37a6d8d5
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>