Extend support for qemu-bridge-helper
The qemu-bridge-helper tool is placed in different locations by
different Linux distributions. Previously, the location of the tool was
hardcoded in the code building the qemu command line needed to boot VM
via QEMU. This patch uses a set of predefined locations where the tool
might reside and looks for the location before building the QEMU command
line. Although still not perfect, the given three locations should cover
most of the cases. In case there will be a need for a different
location, a specific command argument should be added allowing users to
specifcy the location of the qemu-bridge-helper.

Change-Id: I398a7c65da88aac4445419bff835db2bc01a9aa4
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>