Support updating of existing VMs
Image composition is a time consuming task to be repeated on every
change the package manager makes to the underlying files as every time,
image is being composed entire package content along with all their
required packaged are uploaded onto the VM.

This patch introduces a mechanism to cache content that was previously
uploaded. For every file it calculates an MD5 hash value for later
reference. Directories are also hashed, but in this case only the full
VM path is used as directory content. The patch also provides additional
switch to capstan package compose command allowing users to choose
either to compose a new VM or update an existing one.

Verbose switch has also been provided to show informative data to the
user during image composition.

Change-Id: Id9321a84e963a43637db8699f8782f876058801b
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>