Run application package
This patch relieves the user from having to compose an image before
running it. In case capstan run is executed from within an application
package root directory, it will lookup it's package metadata and compose
an image with the name of the package. If the image already existed, it
will only update it with new and modified files and directories.

After the image is composed, it will immediately be executed with the
specified command.

Change-Id: Ia6f4bcf37cf668bf14cc58d5b6485040506e58e3
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>