Check KVM availability when composing a VM
When composing a VM, QEMU is used to launch the OSv launcher with cpiod
daemon waiting to upload content onto the VM. The default command line
to start a VM with QEMU included a switch to enable KVM (-enable-kvm
-cpu host,+x2apic) which prevented it to run in an environment where KVM
is not available (for example, from within a VM without nested
virtualisation).

This patch checks the availability of KVM and sets the switches to the
QEMU command accordingly.

Change-Id: If5f6a152f9516124f38b5a85c64bc65d6559236b
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>