Change meaning of --size argument
Until now the --size argument for 'capstan package compose'
referred to user partition size, excluding loader image size.
Changed it to now refer to total image size since it is otherwise
impossible to guess it. Error is raised when --size is
too small for loader image content to fit in.

Change-Id: I9b92aa84f173c7fffe2e82f856cb043a63534bdf
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>