Automatically pull missing packages
In case the user does not have all the required packages installed in
the local repository, previously capstan would terminate with an error
indicating the missing package. This patch adds support for pulling
packages from the chosen remote repository on demand. By default, this
option is turned off, but can be toggled by passing `--pull-missing` to
the following commands:

 * capstan package compose
 * capstan package collect
 * capstan stack push

Change-Id: I4f31b34f97b1d4106e53e464ae552cc38fc5ea49
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>