Make it possible to disable KVM for QEMU provider
Introduced flag disable_kvm to capstan config.yaml
(or equal env flag CAPSTAN_DISABLE_KVM) which is false
by default. Setting this flag disables HW acceleration.

Usecase to disable KVM is when you want to run another
hypervisor on the same machine as qemu - because KVM
requires exclusive access.

NOTE: When disable_kvm is set to true, both
- capstan package compose
- capstan run
will not use KVM, which might result in performance drop.

Testing:
Before this commit if we started VirtualBox and ran

$ capstan package compose myImg

we would get socket error.
Or if we ran

$ capstan run myImg

we would get 'KVM busy' error.

Starting with this commit, we get additional hint
that disabling KVM might help. When we actually disable
it, the error should dissapear.

Change-Id: If266e4833ee93557c0f92cf67efb8507fad361f8
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>