Create nested folders on collect
Fix bug that 'capstan package compose' command was unable to collect
files in nested folders. The problem was only present when collecting
files from required package tar archive. Root of the problem was that
iterating archived filenames was falsly assumed to be sorted like this:
---
path
path/to
path/to/some
path/to/some/file.txt
---

But in reality, iteration yielded e.g.:
---
path/to/some
path
path/to/some/file.txt
path/to
---

Fixed this by always creating directory structure, not
just the file/link.

Change-Id: I3c725c7eb10ba196021098816c8a376db2feefc2
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>