Switch to maintained version of gophercloud library
Currently, we use github.com/rackspace/gophercloud library to
consume OpenStack REST API with additional PR for Glance support
applied. That forced us to use 'vendor' folder for this library to
simplify capstan installation. But it occured that rackspace/gophercloud
library is no longer maintained.

Fortunately, a successor has appeared at github.com/gophercloud/gophercloud.
So we are switching to this library now to remain up-to-date.

Change-Id: I8ac192b1d687de69f1d530ad092dd17e54633a34
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>