Generate documentation using Capstan --help
We can instruct user to download Capstan and consult its built-in
documentation, but it is more convinient that this documentation
is also part of online documentation. This is especially true when
Capstan is not being used by user directly (e.g. on UniK platform).
Therefore we're introducing a python script to generate .md out of
Capstan inline documentation.

Some things to note:
* script assumes `capstan` executable to exist in $GOPATH/bin
* script assumes to be run from project root directory (e.g.
  $HOME/go/src/github.com/mikelangelo-project/capstan)
* script overrides file ./Documentation/generated/CLI.md
* script should be run manually whenever there's a change is Capstan CLI

Change-Id: I301b799bf77a5759d09c71bdb30557de158d10c0
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>