Remove instance by default
Command `capstan run barney` creates a new instance from image
named 'barney' when we call it for the first time. But when we call
the exact same command for the second time, it just runs instance
named 'barney'. At first glance this sounds reasonable, but it's not -
it caused me many frustrations.

Consider this case:
1 user initializes a new Capstan package
2 user composes image named 'barney' (`capstan package compose barney`)
3 user runs image (`capstan run barney`)
4 CTRL + C ('Woops, I forgot to use bridged network adapter')
5 user runs image again (`capstan run barney -n bridge -b br0`)
6 'Whyyyyyyy do I still get IP of 192.168.122.15 #?$%?#?#?????'
  (answer: because you are just running the same instance as before)

The way it's implemeted before this commit, user should call
`capstan delete barney` prior step 5 to achieve expected behavior, but
I bet she will forget about that!

This commit introduces flag --persist to the `capstan run command`.
When the flag is NOT used, Capstan creates a new instance everytime.
This only applies to qemu provider, behavior for all other providers
is not modified by this commit. This way, step 5 works as user would expect.

Change-Id: Ied18f35e83a38149661896fb8961f36d9ddd1213
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>