Automatically clean orphaned instances
QEMU instances that were not run with --persist flag become
orphaned when we stop them. They have some leftover content in
$HOME/.capstan/instances/qemu/<name> folder. This content has bad
influence on new instance with same name and should therefore be
removed.

This commit implements logic to silently remove orphaned data everytime
user runs `capstan run` command. Note that it is not a problem if we
remove orphaned instance while it is still running - because data on disk
only gets deleted when number of its file descriptors falls to 0.

Change-Id: I5d192cdade23d0b5255e6d7be1423cb85f5b2b18
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>