Allow Capstan to use meta/run.yaml before ignoring it
Having
```
/meta
```
specified as part of built-in capstanignore causes meta folder
from not being iterated. Therefore meta/run.yaml is never processed
and /run/<config-name> files are not generated. Fixed this by updating
the ignore pattern to
```
/meta/*
```
which causes Capstan to ignore all files in /meta, but doesn't prevent
it from iterating the folder.

Change-Id: Ie33016c676df72a7f63bea5567be7936e6495e51
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>