Protect user from ignoring `/meta`
As I was writing a .capstanignore for some Node.js project I
went folder by folder and decide for each whether I want it
ignored or not. And so `/meta` got on my "yeah, I don't need it
in my unikernel" list. BUM! The runscript scripts were not created.

With this commit we protect user from doing the same mistake as I did.

Change-Id: Ifb2b1eecdf91890e9c5236bd73dcbedc5fc4e00a
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>