Use operator `?=` in runscripts to prevent override
Currently Capstan uses `--env=PORT=8000` syntax when generating
runscripts (e.g. /run/master) which overrides environment variables
set to the boot command itself.

This commit replaces the operator `=` with `?=` which results in
runscript syntax like `--env=PORT?=8000`. This operator tells bootloader
to only assign the environment variable if it's not set yet - therefore it
never overrides whatever we set in the bootcmd itself.

Change-Id: I33994543f70208f52228eb7dd6cd67a3da911d9e
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>