Unit test `capstan package describe` command
We add a test where DescribePackage is tested. It turns out that
the function is not testable if it prints directly to console so
we refactor it to print to string. Also, gocheck checker "Matches"
turned out not to be able to match regex agains multiline string
so we had to provide our own MatchesMultiline checker to do the job.

Change-Id: I7096e5eab0dbafb4f4cd21a6e59ff7da781e063c
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>