Write recursive boot commands into /run directory
When collecting the package the meta/run.yaml file gets
parsed and each config_set is stored into a file. Using runscript
utility of OSv kernel is then possible to boot OSv from that file.

With this commit we not only provide /run/<config-set-name> files
for current package, but also for all recursive ones.

Change-Id: Ia340eacd654a926206ad5981a6e596fca4424ee9
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>