Support --env argument for `capstan package compose` command
Similarly as it's already implemented for `capstan run` we want to
be able to accept environment variables here. We need this when we
want to compose unikernel with specific bootcmd and run it elsewhere
(e.g. on openstack).

To support such functionality we do a little refactoring. There were
too many flags walking around the code and with the refactoring we
encapsulate them into a struct named BootOptions. This way we
centralize work related to building boot command. Having this done,
it was simple to support --env also for `package compose` command.

Change-Id: I07f7aa256c290f65adb75539c4e9ecaca8db97aa
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>