Improve readability of unit tests
Since we're adding more and more tests (especially to cmd/package_test.go)
it is vital to keep it as readable as possible. So with this commit we
implement and use following new utility functions (documentation is in code):

* requireFakeDemoPkg
* setRunYaml
* fixIndent

The most importatnt one is fixIndent that enables us to indent multiline yaml
strings so that it matches code flow.

Change-Id: Ifad0c0f628eb32355ec3d49d601d590703ab1203
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>