Rewrite Java runtime to make use of config_set inheritance
Current Java runtime is implemented in a very old fashion that
relies on /etc/javamains file be present on the unikernel. Since
this is now obsolete and brings issues since bootcmd cannot be
set after the unikernel is built, we rewrite the runtime.

We reimplement the GetBootCmd function that now doesn't build the
actual bootcmd on its own anymore, but rather makes use of bootcmd
provided by openjdk8-zulu-compact1 package.

So with this commit we actually transform runtime into a thin wrapper of
a specific package. User could easily use package directly, but the
wrapper offers him some validation and nicer format of arguments.

Change-Id: I6be3e2769788e83d6086c2caadca54c42910b402
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>