Drop unneeded OnCollect function
There was the "OnCollect()" function defined in Runtime interface
that was meant to prepare additional files needed by selected run
configuration. But using this function we actuall lock unikernel during
`capstan package compose` to that specific boot command. In other
words, config set can not be changed when invoking `capstan run` because
no files can be changed at that time.

Luckily, it turns out that there was only Java package using the
problemmatic "OnCollect()" function and that even Java can be used
without it. With this commit we therefore simply remove the "OnCollect()"
function, that only causes problems.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>