Rewrite Node.js runtime to make use of config_set inheritance
Now that node-4.4.5 package contains meta/run.yaml, we are able to
use it also in case of "runtime: node". We ensure that list of available
arguments of "runtime: node" matches those of node-4.4.5 package.

With this commit we also implement support for 'shell' flag in node's
meta/run.yaml to make it ultra simple to run node interactive shell.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>