Marshall empty timestamp into N/A instead of zero timestamp
Currently, if we are marshalling a package with nil value for
'created' attribute, a zero timestamp gets written:

```
created: 0001-01-01 00:00
```

which is wrong. With this commit we rather marshall nil value into
a N/A:

```
created: N/A
```

so that unmarshalling results into a nil date again, and not
the zero one.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>