Let ListImages() return a string instead printing to stdout
With this commit we modify ListImages so that it now returns
a string, which opens an opportunity to provide unit tests
for it. We grab the opportunity and actually provide the unit
tests. Having tests in place, we're confident enough to
refactor the body of ListImages method that was implemented
in a strange way and thus hardly understandable.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>