Modernize Java runtime
An issue opened by a user trying to run .jar file on OSv pointed out
that some more documentation should be provided for each runtime. While
we've done this for Node.js already, we're now providing it for Java
runtime as well. While providing documentation it turned out that we're
actually lacking a proper way to support switching Java package underneath.
So with this commit we support this as well.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>