Attach volume on `capstan run`
Until now we were not able to attach volume when using Capstan.
With this commit we bring basic support for it by introducing
following parameter to the command:

```
$ capstan run demo --volume ./test1.img --volume ./volume2.img --volume ./volume3.img
```

At the moment we only support volumes for QEMU hypervisor and print
a warning if --volume option is used with any other hypervisor.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>