Create volumes with Capstan
With this commit we implement a simple wrapper around qemu-img to
simplify volume creation for user by using Capstan:

```
$ capstan volume create vol1 --size 1GB --format qcow2
```

The above command needs to be executed inside package (presence of meta/package.yaml
is tested) and will create `volumes` directory with two files:

```
<package_root>
  |-volumes
    |-vol1      # volume itself
    |-vol1.yaml # metadata for Capstan
```

The metadata file is currently not used by Capstan yet, but will be soon
since Capstan will automatically detect all the volumes inside ./volumes
directory. Also `capstan volume delete` is implemented in this commit and
`./volumes` directory is implicitly added to the .capstanignore.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>