Use QEMU aio=threads by default
Currently we use aio=native by default and then conditionally
swap it to aio=threads depending on the platform where run. But
it seems like that support for "native" is going away. On Ubuntu
16 it still works, but on Ubuntu 17 it doesn't work anymore. With
this commit we therefore use a more general option by default and
allow user to explicitly set "native" in $HOME/.capstan/config.yaml
when desired.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>