Don't require osv.bootstrap when compose-remote
With this commit we implement a logic that decides which
packages will be implicitly required when collecting files.
In case of `capstan package compose` we require osv.bootstrap
while in case of `capstan package compose-remote` we require
osv.compose-remote.

Why? When cpiod is running in remote unikernel, we MUST NOT upload
the /tools/cpiod.so binary again or it crashes. Therefore we're
better of if we omit uploading osv.bootstrap package at all.
Instead we rather implicitly require the osv.compose-remote to
baby-sit user even further (besides omitting the bootstrap).

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>